Vinuta Assignment/Question 2 - Django Application Development/django_rest/requirement.txt
Django==3.1.2
djangorestframework==3.12.1
pytz==2019.2
tzlocal==2.0.0